utils clean up
